Start pulling out a core module from the implementation module
Higher up socket will also delegate to the core module directly for
`{:core, message}` casts.
